Implement search for events
On each calendar page an input box for the event search is displayed.
The search is limited to the calendar shown.
All events occurring up to one month away from the day selected in the calendar are taken into account.
Events that occur in the past are emphasized.

In the result list the start date and end date of the respective events are clickable.
Selecting one switches to the appropriate date in the calendar.

When a repeated event is created an additional check is performed to prevent the creation of events that overlap with itself in the future.

	modified:   calendar/app/controllers/Application.java
	modified:   calendar/app/models/Calendar.java
	new file:   calendar/app/views/Application/searchEvent.html
	modified:   calendar/app/views/Application/showCalendar.html